backend/pyproject.toml
[project]
name = "intake-iq"
version = "0.1.0"
description = "AI-powered insurance claim intake assistant"
authors = [
    { name = "Viacheslav Danilov", email = "viacheslav.v.danilov@gmail.com" },
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "azure-ai-documentintelligence==1.0.0b4",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "openai>=2.30.0",
    "pre-commit>=4.5.1",
    "pytest>=8.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/intake_iq"]

[tool.mypy]
python_version = "3.13"
exclude = ["venv", ".venv", "tests"]
ignore_missing_imports = true
disallow_untyped_decorators = false
disallow_untyped_calls = false
explicit_package_bases = true

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["intake_iq"]
